requirements.txt
beautifultable==0.8.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
fandogh-cli==1.26.10
idna==3.3
Pillow==9.1.1
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2022.1
PyYAML==5.4.1
requests==2.28.1
requests-toolbelt==0.9.1
retrying==1.3.3
rsa==4.8
six==1.16.0
Telethon==1.24.0
tzlocal==2.1
urllib3==1.26.9
websocket-client==1.3.3
